requirements.txt
Django==4.2.5
psycopg2-binary==2.9.9
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0
django-cors-headers==4.1.0
requests==2.31.0
stripe==3.4.0
openai==1.57.2
djangorestframework==3.15.2